Add finance_calculator.py updated

+++ python_introduction/finance_calculator.py
@@ -1,8 +1,8 @@
 #Personal Finance Calculator
 
 #prompt the user to enter a monthly income and the total monthly expenses
-monthly_income = int(input("Enter your monthly income: "))
-total_monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 
 #calculate monthly savings
 monthly_savings = monthly_income - total_monthly_expenses
